refact: move root saga and reducer files
