--- src/reducers/index.ts
@@ -1,19 +1,0 @@
-export { default as app }                  from './app';
-export { default as auth }                 from 'services/auth/reducer.js';
-export { default as flashMessages }        from 'services/flash-messages/reducer.js';
-export { default as invoice }              from 'services/invoices/entity-reducer.js';
-export { default as invoices }             from 'services/invoices/entities-reducer.js';
-export { default as unpaidInvoices }       from 'services/invoices/unpaid-reducer.js';
-export { default as unpaidInvoicesCount }  from 'services/invoices/unpaid-count-reducer.js';
-export { default as organizations }        from 'services/organizations/entities-reducer.js';
-export { default as currentOrganization }  from 'services/organizations/current-reducer.js';
-export { default as transaction }          from 'services/transactions/entity-reducer.js';
-export { default as transactions }         from 'services/transactions/entities-reducer.js';
-export { default as members }              from 'services/members/reducer.js';
-export { default as bankAccounts }         from 'services/bank-accounts/entities-reducer.js';
-export { default as bankAccount }          from 'services/bank-accounts/entity-reducer.js';
-export { default as categories }           from 'services/categories/entities-reducer.js';
-export { default as category }             from 'services/categories/entity-reducer.js';
-export { default as customers }            from 'services/customers/entities-reducer.js';
-export { default as customer }             from 'services/customers/entity-reducer.js';
-export { default as balances }             from 'services/balances/reducer.js';
