+++ src/services/reducers.js
@@ -1,0 +1,19 @@
+export { default as app }                  from './app/reducer.js';
+export { default as auth }                 from './auth/reducer.js';
+export { default as balances }             from './balances/reducer.js';
+export { default as bankAccounts }         from './bank-accounts/entities-reducer.js';
+export { default as bankAccount }          from './bank-accounts/entity-reducer.js';
+export { default as categories }           from './categories/entities-reducer.js';
+export { default as category }             from './categories/entity-reducer.js';
+export { default as currentOrganization }  from './organizations/current-reducer.js';
+export { default as customers }            from './customers/entities-reducer.js';
+export { default as customer }             from './customers/entity-reducer.js';
+export { default as flashMessages }        from './flash-messages/reducer.js';
+export { default as invoice }              from './invoices/entity-reducer.js';
+export { default as invoices }             from './invoices/entities-reducer.js';
+export { default as members }              from './members/reducer.js';
+export { default as organizations }        from './organizations/entities-reducer.js';
+export { default as transaction }          from './transactions/entity-reducer.js';
+export { default as transactions }         from './transactions/entities-reducer.js';
+export { default as unpaidInvoices }       from './invoices/unpaid-reducer.js';
+export { default as unpaidInvoicesCount }  from './invoices/unpaid-count-reducer.js';
